[OM] Fix finalization of nested ReferenceValue
After instantiating an Object in the Evaluator, we recursively finalize its value
to eliminate intermediate ReferenceValues, ensuring that the original type of
om.any is preserved. This patch fixes an issue where, if a ReferenceValue
remains a ReferenceValue after being stripped, the finalization process should
continue but previously did not (e.g., a list of om.any that is an object
returns another list of om.any that is also an object).

Signed-off-by: unlsycn <unlsycn@unlsycn.com>